train val scripts/requirements.txt
# 基础依赖
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0

# 深度学习框架
torch>=1.9.0
torchvision>=0.10.0

# 图像处理和数据增强
Pillow>=8.3.0
albumentations>=1.0.3

# 日志记录和实验跟踪
mlflow>=1.20.0